test(checkbox): moved test files into seperate folder
