+++ packages/core/src/components/checkbox/test/checkbox.e2e.ts
